services/colpali_service/requirements.txt
colpali-engine
fastapi
uvicorn[standard]
pillow
numpy
pydantic
requests
hf_xet